Close UDP socket after droping client
